agentformer/requirements.txt
flask
python-dotenv
openai
sentence-transformers
numpy 